feat(contact-tagging): Backend: Tag filter handlers unified
